Add some documentation/cleanup
