+++ .gitignore
@@ -12,11 +12,16 @@
 **/.pytest_cache/
 
 ## VSCode .gitignore (included from https://github.com/github/gitignore/blob/master/Global/VisualStudioCode.gitignore
-.vscode/
+.vscode/*
 !.vscode/settings.json
 !.vscode/tasks.json
 !.vscode/launch.json
 !.vscode/extensions.json
+!.vscode/*.code-snippets
+!*.code-workspace
+
+# Built Visual Studio Code Extensions
+*.vsix
 
 ## Ignore Visual Studio temporary files, build results, and
 ## files generated by popular Visual Studio add-ons.
