+++ scripts/run_performance_job.py
@@ -1,5 +1,4 @@
 from logging import getLogger
-from pathlib import Path
 import re
 from dataclasses import dataclass, field
 from datetime import timedelta
@@ -9,16 +8,15 @@ import os
 import shutil
 from subprocess import CalledProcessError
 import sys
-import tarfile
 import tempfile
 from traceback import format_exc
 import urllib.request
 import xml.etree.ElementTree as ET
 from typing import Any, Dict, List, Optional
-import zipfile
 
+from build_runtime_payload import *
 import ci_setup
-from performance.common import RunCommand, iswin, set_environment_variable
+from performance.common import RunCommand, set_environment_variable
 from performance.logger import setup_loggers
 from send_to_helix import PerfSendToHelixArgs, perf_send_to_helix
 
@@ -320,199 +318,6 @@ def logical_machine_to_queue(logical_machine: str, internal: bool, os_group: str
             }
             return queue_map.get(logical_machine, "Ubuntu.2204.Amd64.Tiger.Perf")
 
-def extract_archive_or_copy(archive_path_or_dir: str, dest_dir: str, prefix: Optional[str] = None):
-    if not os.path.exists(archive_path_or_dir):
-        raise FileNotFoundError(f"Archive or directory not found: {archive_path_or_dir}")
-    
-    if os.path.exists(dest_dir) and not os.path.isdir(dest_dir):
-        raise NotADirectoryError(f"Destination exists and is not a directory: {dest_dir}")
-    
-    os.makedirs(dest_dir, exist_ok=True)
-    # get everything up to last slash
-    prefix_folder = prefix or ""
-    if prefix and not prefix.endswith("/") and "/" in prefix:
-        prefix_folder = prefix[:prefix.rfind("/") + 1]
-
-    if os.path.isdir(archive_path_or_dir):
-        src_dir = archive_path_or_dir
-        if prefix is not None:
-            src_dir = os.path.join(archive_path_or_dir, prefix_folder)
-            if not os.path.exists(src_dir):
-                raise FileNotFoundError(f"Source folder not found in archive: {src_dir}")
-            prefix = prefix[len(prefix_folder):]
-
-        if not os.path.samefile(src_dir, dest_dir):
-            if not prefix:
-                shutil.copytree(src_dir, dest_dir, dirs_exist_ok=True)
-            else:
-                for item in Path(src_dir).rglob(f"{prefix}*"):
-                    if item.is_file():
-                        dest_path = os.path.join(dest_dir, item.relative_to(src_dir))
-                        os.makedirs(os.path.dirname(dest_path), exist_ok=True)
-                        shutil.copy2(item, dest_path)
-    elif archive_path_or_dir.endswith(".zip"):
-        with zipfile.ZipFile(archive_path_or_dir, 'r') as zip_ref:
-            if prefix is None:
-                zip_ref.extractall(dest_dir)
-            else:
-                for member in zip_ref.namelist():
-                    if member.startswith(prefix):
-                        relative_path = member[len(prefix_folder):]
-                        if not relative_path or relative_path.endswith("/"):
-                            continue
-                        output_path = os.path.join(dest_dir, relative_path)
-                        os.makedirs(os.path.dirname(output_path), exist_ok=True)
-                        with zip_ref.open(member) as source, open(output_path, 'wb') as target:
-                            target.write(source.read())
-    elif archive_path_or_dir.endswith(".tar.gz"):
-        with tarfile.open(archive_path_or_dir, 'r:gz') as tar_ref:
-            if prefix is None:
-                tar_ref.extractall(dest_dir)
-            else:
-                for member in tar_ref.getmembers():
-                    if member.name.startswith(prefix):
-                        relative_path = member.name[len(prefix_folder):]
-                        if not relative_path or relative_path.endswith("/"):
-                            continue
-                        output_path = os.path.join(dest_dir, relative_path)
-                        os.makedirs(os.path.dirname(output_path), exist_ok=True)
-                        source = tar_ref.extractfile(member)
-                        if source is not None:
-                            with source and open(output_path, 'wb') as target:
-                                target.write(source.read())
-    else:
-        raise Exception("Unsupported archive format")
-
-def build_coreclr_core_root(
-        runtime_repo_dir: str,
-        core_root_dest: str,
-        os_group: str,
-        architecture: str,
-        coreclr_archive_or_dir: Optional[str] = None,
-        libraries_config: Optional[str] = None,
-        clean_artifacts: bool = False):
-    if not os.path.exists(runtime_repo_dir):
-        raise Exception("Runtime repo directory not found")
-    
-    if iswin() and os_group != "windows":
-        raise Exception(f"Unable to build Core_Root for {os_group} on Windows")
-
-    artifacts_dir = os.path.join(runtime_repo_dir, "artifacts")
-    artifacts_bin_dir = os.path.join(artifacts_dir, "bin")
-    artifacts_tests_dir = os.path.join(artifacts_dir, "tests")
-    
-    coreclr_archive_or_dir = coreclr_archive_or_dir or artifacts_bin_dir # default to bin dir
-    if not os.path.exists(coreclr_archive_or_dir):
-        raise Exception("CoreCLR build not found")
-    
-    if clean_artifacts:
-        if os.path.exists(artifacts_bin_dir) and not os.path.samefile(coreclr_archive_or_dir, artifacts_bin_dir):
-            shutil.rmtree(artifacts_bin_dir)
-        if os.path.exists(artifacts_tests_dir):
-            shutil.rmtree(artifacts_tests_dir)
-
-    extract_archive_or_copy(coreclr_archive_or_dir, artifacts_bin_dir)
-    
-    build_file = "build.cmd" if iswin() else "build.sh"
-    build_script = os.path.join(runtime_repo_dir, "src", "tests", build_file)
-    if not os.path.exists(build_script):
-        raise Exception(f"Build script not found at path: {build_script}")
-
-    generate_layout_command = [build_script, "release", architecture, "generatelayoutonly"]
-
-    if not iswin():
-        generate_layout_command.extend(["-os", os_group])
-
-    if libraries_config:
-        generate_layout_command.append(f"/p:LibrariesConfiguration={libraries_config}")
-
-    getLogger().info(f"Generating Core_Root")
-    RunCommand(generate_layout_command, verbose=True).run(runtime_repo_dir)
-
-    core_root_dir = os.path.join(artifacts_tests_dir, "coreclr", f"{os_group}.{architecture}.Release", "Tests", "Core_Root")
-    if not os.path.exists(core_root_dir):
-        raise Exception(f"Core_Root directory not found in expected location: {core_root_dir}")
-    
-    shutil.copytree(core_root_dir, core_root_dest, dirs_exist_ok=True, ignore=shutil.ignore_patterns("*.pdb"))
-
-def build_mono_payload(
-        mono_payload_dst: str,
-        os_group: str,
-        framework: str,
-        build_config: str,
-        architecture: str,
-        product_version: str,
-        runtime_repo_dir: Optional[str] = None,
-        mono_archive_or_dir: Optional[str] = None):
-                
-    if mono_archive_or_dir is None:
-        if runtime_repo_dir is None:
-            raise Exception("Please provide a path to the built mono artifacts")
-        mono_archive_or_dir = os.path.join(runtime_repo_dir, "artifacts", "bin")
-
-    build_config_upper = build_config.title() # release -> Release
-
-    extract_archive_or_copy(mono_archive_or_dir, mono_payload_dst, prefix=f"testhost/{framework}-{os_group}-{build_config_upper}-{architecture}/")
-    corerun_target = os.path.join(mono_payload_dst, "shared", "Microsoft.NETCore.App", product_version)
-    extract_archive_or_copy(mono_archive_or_dir, corerun_target, prefix=f"coreclr/{os_group}.{architecture}.{build_config_upper}/corerun") # no trailing slash as we want files starting with corerun
-
-def build_monoaot_payload(
-        monoaot_artifacts_archive_or_dir: str,
-        payload_dest: str,
-        architecture: str):
-    
-    pack_dir = os.path.join(payload_dest, "pack")
-    os.makedirs(pack_dir, exist_ok=True)
-
-    extract_archive_or_copy(
-        monoaot_artifacts_archive_or_dir, 
-        payload_dest, 
-        prefix=f"artifacts/bin/mono/linux.{architecture}.Release/cross/linux-{architecture}/")
-    
-    extract_archive_or_copy(
-        monoaot_artifacts_archive_or_dir, 
-        pack_dir, 
-        prefix=f"artifacts/bin/microsoft.netcore.app.runtime.linux-{architecture}/Release/")
-    
-def build_wasm_payload(
-        browser_wasm_archive_or_dir: str,
-        payload_parent_dir: str, # wasm creates three payload directories
-        test_main_js_path: Optional[str] = None,
-        runtime_repo_dir: Optional[str] = None):
-    
-    if test_main_js_path is None:
-        if runtime_repo_dir is None:
-            raise Exception("Please provide a path to the test-main.js or runtime repository")
-        test_main_js_path = runtime_repo_dir and os.path.join(runtime_repo_dir, "src", "mono", "browser", "test-main.js")
-    
-    if not os.path.exists(test_main_js_path):
-        raise Exception(f"test-main.js not found in expected location: {test_main_js_path}")
-    
-    wasm_dotnet_dir = os.path.join(payload_parent_dir, "dotnet")
-    wasm_built_nugets_dir = os.path.join(payload_parent_dir, "built-nugets")
-    wasm_data_dir = os.path.join(payload_parent_dir, "wasm-data")
-
-    extract_archive_or_copy(
-        browser_wasm_archive_or_dir,
-        wasm_dotnet_dir,
-        prefix="staging/dotnet-latest/")
-    
-    extract_archive_or_copy(
-        browser_wasm_archive_or_dir,
-        wasm_built_nugets_dir,
-        prefix="staging/built-nugets/")
-
-    os.makedirs(wasm_data_dir, exist_ok=True)
-    shutil.copy(test_main_js_path, os.path.join(wasm_data_dir, "test-main.js"))
-
-    for dir in [wasm_dotnet_dir, wasm_built_nugets_dir, wasm_data_dir]:
-        for root, _, files in os.walk(dir):
-            for item in files:
-                try:
-                    os.chmod(os.path.join(root, item), 0o664) # rw-rw-r--
-                except OSError as e:
-                    print(f"Failed to set permissions for {item}: {e}")
-
 def get_bdn_arguments(
         run_categories: str,
         internal: bool,
@@ -772,20 +577,12 @@ def run_performance_job(args: RunPerformanceJobArgs):
     
     ios_mono = args.runtime_type == "iOSMono"
     ios_nativeaot = args.runtime_type == "iOSNativeAOT"
-    mono_aot = False
-    mono_dotnet = False
-    wasm = False
-    wasm_aot = False
-    product_version = None
-    if args.runtime_type == "mono":
-        if args.codegen_type.lower() == "aot":
-            mono_aot = True
-        else:
-            mono_dotnet = True
-    elif args.runtime_type == "wasm":
-        wasm = True
-        if args.codegen_type.lower() == "aot":
-            wasm_aot = True
+    is_aot = args.codegen_type.lower() == "aot"
+    is_mono = args.runtime_type == "mono"
+    mono_aot = is_mono and is_aot
+    mono_dotnet = is_mono and not is_aot
+    wasm = args.runtime_type == "wasm"
+    wasm_aot = wasm and is_aot
 
     working_dir = os.path.join(args.performance_repo_dir, "CorrelationStaging") # folder in which the payload and workitem directories will be made
     work_item_dir = os.path.join(working_dir, "workitem", "") # Folder in which the work item commands will be run in
@@ -873,6 +670,7 @@ def run_performance_job(args: RunPerformanceJobArgs):
     if not args.internal and not args.performance_repo_ci:
         ci_setup_arguments.not_in_lab = True
 
+    product_version = None
     if mono_dotnet and not mono_aot:
         if args.framework is None:
             raise Exception("Framework must be specified for Mono dotnet runs")
@@ -993,7 +791,7 @@ def run_performance_job(args: RunPerformanceJobArgs):
             if args.runtime_repo_dir is None:
                 raise Exception("Core_Root directory must be specified for non-performance CI runs")
             
-            build_coreclr_core_root(
+            build_coreroot_payload(
                 args.runtime_repo_dir, 
                 core_root_dest=coreroot_payload_dir, 
                 os_group=args.os_group,
